\name{emission_hessian_louis}
\alias{emission_hessian_louis}
\title{emission_hessian_louis
Get the complete data hessian matrix for emission parameters}
\usage{
  emission_hessian_louis(emission.setup, param.values,
    emission.counts)

  emission_hessian_louis(emission.setup, param.values,
    emission.counts)
}
\arguments{
  \item{emission.setup}{list emission setup object}

  \item{param.values}{values of emission parameters}

  \item{emission.counts}{array (dim num states x num obs
  states x num subjects) with the (expected) counts of the
  o_ij}

  \item{emission.setup}{list emission setup object}

  \item{param.values}{values of emission parameters}

  \item{emission.counts}{array (dim num states x num obs
  states x num subjects) with the (expected) counts of the
  o_ij}
}
\value{
  complete data hessian matrix for emission parameters

  complete data hessian matrix for emission parameters
}
\description{
  emission_hessian_louis Get the complete data hessian
  matrix for emission parameters

  emission_hessian_louis Get the complete data hessian
  matrix for emission parameters
}

