\name{get.deriv.array}
\alias{get.deriv.array}
\title{Get deriv arrays based on the non-fixed parameters}
\usage{
  get.deriv.array(covariate.array, param.types)
}
\arguments{
  \item{param.types}{(not fixed if=0)}

  \item{covariate.array}{}
}
\description{
  Get deriv arrays based on the non-fixed parameters
}

