\name{init_hessian_louis}
\alias{init_hessian_louis}
\title{Get the complete data hessian matrix for initial dist parameters}
\usage{
  init_hessian_louis(init.setup, param.values)

  init_hessian_louis(init.setup, param.values)
}
\arguments{
  \item{init.setup}{list init setup object}

  \item{param.values}{values of init parameters}

  \item{init.setup}{list init setup object}

  \item{param.values}{values of init parameters}
}
\value{
  complete data hessian matrix for init parameters

  complete data hessian matrix for init parameters
}
\description{
  Get the complete data hessian matrix for initial dist
  parameters

  Get the complete data hessian matrix for initial dist
  parameters
}

