\name{rate_hessian_louis}
\alias{rate_hessian_louis}
\title{rate_hessian_louis
Get the complete data hessian matrix for rate parameters}
\usage{
  rate_hessian_louis(rates.setup, rate.param.values,
    durations)

  rate_hessian_louis(rates.setup, rate.param.values,
    durations)
}
\arguments{
  \item{rates.setup}{list rate setup object}

  \item{rate.param.values}{values of rate parameters}

  \item{durations}{array (dim num.states x num.subjects)
  with the expected durations}

  \item{rates.setup}{list rate setup object}

  \item{rate.param.values}{values of rate parameters}

  \item{durations}{array (dim num.states x num.subjects)
  with the expected durations}
}
\value{
  complete data hessian matrix for rate parameters

  complete data hessian matrix for rate parameters
}
\description{
  rate_hessian_louis Get the complete data hessian matrix
  for rate parameters

  rate_hessian_louis Get the complete data hessian matrix
  for rate parameters
}

