\name{s_1_Ni}
\alias{s_1_Ni}
\title{s function for transitions s_k(x_k,x_{k+1})[1:num.times,stat[l]], if the type of stat [l] is Ni}
\usage{
  s_1_Ni(Ni, num_states, num_times)
}
\arguments{
  \item{Ni}{hidden state i}

  \item{num_states}{}

  \item{num_times}{}
}
\value{
  s array of dim(num_states,num_states,num.times,1)
}
\description{
  s function for transitions
  s_k(x_k,x_{k+1})[1:num.times,stat[l]], if the type of
  stat [l] is Ni
}

