\name{s_1_zi}
\alias{s_1_zi}
\title{s function for transitions s_k(x_k,x_{k+1})[1:num.times,stat[l]], if the type of stat [l] is Z_i}
\usage{
  s_1_zi(zi, num_states, num_times)
}
\arguments{
  \item{zi}{state i}

  \item{num_states}{}

  \item{num_times}{}
}
\value{
  s array of dim(num_states,num_states,num.times,1)
}
\description{
  s function for transitions
  s_k(x_k,x_{k+1})[1:num.times,stat[l]], if the type of
  stat [l] is Z_i
}

